fix EdgeIndex size. use uint64_t

+++ core/common/types.h
@@ -14,7 +14,7 @@ typedef uint32_t VertexID;
 typedef uint32_t VertexIndex;
 typedef uint32_t VertexLabel;
 typedef uint32_t VertexCount;
-typedef uint32_t EdgeIndex;
+typedef uint64_t EdgeIndex;
 
 typedef uint8_t DefaultEdgeDataType;  // used for position
 typedef uint32_t Uint32VertexDataType;
